@@ -1,4 +1,3 @@
 const privateKey = 'abC123!';
 
-// node & browser
-module.exports = privateKey;
+export default privateKey;
