@@ -1,7 +1,7 @@
 /**
  * JWT
  */
-export const JWT_SECRET: "abC123!";
+export const JWT_SECRET: any;
 export const JWT_EXPIRES_IN: number;
 export function sign(payload: any, privateKey: any, header: any): string;
 export function decode(token: any): any;
